Dev03

+++ src/main/java/com/restapi/ecommerce/service/ProductServiceImpl.java
@@ -147,7 +147,7 @@ public class ProductServiceImpl implements ProductService {
 						-> new ResourceNotFoundException("Product", "productId", prodId));
 		String imageName = productToDelete.getImageName();
 		// if there's an image file, delete it from S3 bucket. 
-		if (!(imageName == "") || !(imageName == null)) {
+		if ((imageName != "") && (imageName != null)) {
 			imgUploadService.deleteImg(imageName);
 		}
 	    productToDelete.setDeletedAt(Instant.now());
